backend/requirements.txt
Flask
python-dotenv
Flask-PyMongo
PyMongo
pyJWT
sqlalchemy